requirements.txt
discord.py==2.6.3
python-dotenv==1.0.0
requests==2.31.0
fuzzywuzzy==0.18.0
flask==2.3.2